Pull Tidy Data from LERS

Pull Tidy Data from LERS

get_enrolment(measures, rows, institutions, username, password, print = F,
  remove.offshores = T, remove.continuingstudies = T, postalcodes,
  censusdivisions, cipcodes, sa.mh = F)

Arguments

measures
Select possible measurements

rows
Select rows to cut data by

institutions
Optional: Filters by Institution. Can exclude by including "exclude" in the vector, then the items in that vector will be excluded rather than included.

username
Optional: Either supply a siams username or use .Rprofile otherwise as "siams.username".

password
Optional: Either supply a siams password or use .Rprofile otherwise as "siams.password".

print
Optional: Prints the MDX string used to query the database.

postalcodes
Optional: Filters by Postal Code. Can exclude by including "exclude" in the vector, then the items in that vector will be excluded rather than included.

censusdivisions
Optional: Filters by Census Division. Can exclude by including "exclude" in the vector, then the items in that vector will be excluded rather than included.

cipcodes
Optional: Filters by Cip Code of any length, can used mixed vector with any length cip codes together.

Examples

df <- get_enrolment(measures = c("Unique Student Static", "FLE"),
                rows = c("Provider", "Gender"),
                institutions = c("MH", "MU", "UA"),
                postalcodes = c("exclude", "T1C"),
                censusdivision = c("exclude", "1"))